backend/requirements.txt
anyio==3.4.0
asgiref==3.4.1
click==8.0.3
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
pkg_resources==0.0.0
pydantic==1.8.2
PyYAML==6.0
sniffio==1.2.0
SQLAlchemy==1.4.27
starlette==0.16.0
typing_extensions==4.0.1
uvicorn==0.15.0
